requirements.txt
matplotlib==3.2.2
nltk==3.5
wordcloud==1.7.0
pandas==1.0.5
numpy==1.18.5
squarify==0.4.3
PyPDF2==1.26.0
fasttext
GetOldTweets3==0.0.11
langdetect
googletrans==3.0.0
Pillow
flask==1.1.2
sklearn
transformers==3.0.2
seaborn==0.10.1
keras==2.0.8
lemmagen
tensorflow_hub==0.8.0
torch==1.5.0
tensorflow==1.14.0
